Add initial version after one day of hacking
This is not secure at all, which should be reconsidered.
Also, it does not check for escaping properly yet, and
is not finished past the 'adding of new options' stage.

The most significant next todo is to figure out how to
best set timers to proceed to next steps automatically.